+++ content_tracks.go
@@ -16,8 +16,11 @@ import (
 	"sort"
 	"strconv"
 	"strings"
+	"sync"
+	"time"
 
 	"justapengu.in/acsm/cmd/server-manager/static"
+	"justapengu.in/acsm/internal/acserver"
 	"justapengu.in/acsm/pkg/ai"
 
 	"github.com/cj123/ini"
@@ -342,6 +345,19 @@ func (th *TracksHandler) trackImage(w http.ResponseWriter, r *http.Request) {
 	} else {
 		w.Header().Add("Content-Length", strconv.Itoa(int(n)))
 	}
+}
+
+func (th *TracksHandler) rebuildTrackMaps(w http.ResponseWriter, r *http.Request) {
+	go panicCapture(func() {
+		err := th.trackManager.RebuildAllTrackMaps()
+
+		if err != nil {
+			logrus.WithError(err).Error("could not rebuild track maps")
+		}
+	})
+
+	AddFlash(w, r, "Started re-building track maps! This may take some time.")
+	http.Redirect(w, r, r.Referer(), http.StatusFound)
 }
 
 func (th *TracksHandler) trackSplineImage(w http.ResponseWriter, r *http.Request) {
@@ -802,14 +818,95 @@ func (tm *TrackManager) UpdateTrackMetadata(name string, r *http.Request) error 
 	return track.MetaData.Save(name)
 }
 
+func (tm *TrackManager) BuildTrackMap(track, layout string) error {
+	trackPath := filepath.Join(ServerInstallPath, "content", "tracks", track)
+
+	if layout != "" {
+		trackPath = filepath.Join(trackPath, layout)
+	}
+
+	fastLaneSpline, err := ai.ReadSpline(filepath.Join(trackPath, "ai", "fast_lane.ai"))
+
+	if os.IsNotExist(err) {
+		logrus.Debugf("Cannot build track map for %s (%s), needs fast_lane.ai", track, layout)
+		return nil
+	} else if err != nil {
+		return err
+	}
+
+	fullPitLane, err := ai.ReadSpline(filepath.Join(trackPath, "ai", "pit_lane.ai"))
+
+	if os.IsNotExist(err) {
+		logrus.Debugf("Cannot build track map for %s (%s), needs pit_lane.ai", track, layout)
+		return nil
+	} else if err != nil {
+		return err
+	}
+
+	drsZones, _ := acserver.LoadDRSZones(filepath.Join(trackPath, "data", drsZonesFilename))
+
+	renderer := ai.NewTrackMapRenderer(fastLaneSpline, fullPitLane, drsZones)
+
+	mapPNG, err := os.Create(filepath.Join(trackPath, "map.png"))
+
+	if err != nil {
+		return err
+	}
+
+	defer mapPNG.Close()
+
+	trackMapData, err := renderer.Render(mapPNG)
+
+	if err != nil {
+		return err
+	}
+
+	return trackMapData.Save(filepath.Join(trackPath, "data", "map.ini"))
+}
+
+var trackMapRebuildMutex sync.Mutex
+
+func (tm *TrackManager) RebuildAllTrackMaps() error {
+	trackMapRebuildMutex.Lock()
+	defer trackMapRebuildMutex.Unlock()
+
+	logrus.Infof("Building Track Maps for all Tracks")
+	started := time.Now()
+
+	tracks, err := tm.ListTracks()
+
+	if err != nil {
+		return err
+	}
+
+	for _, track := range tracks {
+		for _, layout := range track.Layouts {
+			if layout == defaultLayoutName {
+				layout = ""
+			}
+
+			if err := tm.BuildTrackMap(track.Name, layout); err != nil {
+				logrus.WithError(err).Errorf("Could not build track map for: %s (%s)", track.Name, layout)
+				continue
+			} else {
+				logrus.Infof("Rebuilt track map for: %s (%s)", track.Name, layout)
+			}
+		}
+	}
+
+	logrus.Infof("Track Map build is complete (took: %s)", time.Since(started).String())
+
+	return nil
+}
+
 type TrackDataGateway interface {
 	TrackInfo(name, layout string) (*TrackInfo, error)
-	TrackMap(name, layout string) (*TrackMapData, error)
+	TrackMap(name, layout string) (*ai.TrackMapData, error)
 }
 
 type filesystemTrackData struct{}
 
-func (filesystemTrackData) TrackMap(name, layout string) (*TrackMapData, error) {
+func (filesystemTrackData) TrackMap(name, layout string) (*ai.TrackMapData, error) {
 	return LoadTrackMapData(name, layout)
 }
 
@@ -829,17 +926,7 @@ func (filesystemTrackData) TrackInfo(name, layout string) (*TrackInfo, error) {
 	return trackInfo, err
 }
 
-type TrackMapData struct {
-	Width       float64 `ini:"WIDTH" json:"width"`
-	Height      float64 `ini:"HEIGHT" json:"height"`
-	Margin      float64 `ini:"MARGIN" json:"margin"`
-	ScaleFactor float64 `ini:"SCALE_FACTOR" json:"scale_factor"`
-	OffsetX     float64 `ini:"X_OFFSET" json:"offset_x"`
-	OffsetZ     float64 `ini:"Z_OFFSET" json:"offset_y"`
-	DrawingSize float64 `ini:"DRAWING_SIZE" json:"drawing_size"`
-}
-
-func LoadTrackMapData(track, trackLayout string) (*TrackMapData, error) {
+func LoadTrackMapData(track, trackLayout string) (*ai.TrackMapData, error) {
 	p := filepath.Join(ServerInstallPath, "content", "tracks", track)
 
 	if trackLayout != "" {
@@ -868,7 +955,7 @@ func LoadTrackMapData(track, trackLayout string) (*TrackMapData, error) {
 		return nil, err
 	}
 
-	var mapData TrackMapData
+	var mapData ai.TrackMapData
 
 	if err := s.MapTo(&mapData); err != nil {
 		return nil, err