+++ content_upload.go
@@ -29,34 +29,41 @@ type ContentFile struct {
 	Size     int    `json:"size"`
 }
 
+type UploadPayload struct {
+	Files             []ContentFile `json:"Files"`
+	GenerateTrackMaps bool          `json:"GenerateTrackMaps"`
+}
+
 var base64HeaderRegex = regexp.MustCompile("^(data:.+;base64,)")
 
 type ContentUploadHandler struct {
 	*BaseHandler
 
-	carManager *CarManager
+	carManager   *CarManager
+	trackManager *TrackManager
 }
 
-func NewContentUploadHandler(baseHandler *BaseHandler, carManager *CarManager) *ContentUploadHandler {
+func NewContentUploadHandler(baseHandler *BaseHandler, carManager *CarManager, trackManager *TrackManager) *ContentUploadHandler {
 	return &ContentUploadHandler{
-		BaseHandler: baseHandler,
-		carManager:  carManager,
+		BaseHandler:  baseHandler,
+		carManager:   carManager,
+		trackManager: trackManager,
 	}
 }
 
 // Stores Files encoded into r.Body
 func (cuh *ContentUploadHandler) upload(contentType ContentType) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		var files []ContentFile
+		var uploadPayload UploadPayload
 
-		err := json.NewDecoder(r.Body).Decode(&files)
+		err := json.NewDecoder(r.Body).Decode(&uploadPayload)
 
 		if err != nil {
 			logrus.WithError(err).Errorf("could not decode %s json", contentType)
 			return
 		}
 
-		err = cuh.addFiles(files, contentType)
+		err = cuh.handleUploadPayload(uploadPayload, contentType)
 
 		if err != nil {
 			logrus.WithError(err).Error("couldn't upload file")
@@ -68,8 +75,7 @@ func (cuh *ContentUploadHandler) upload(contentType ContentType) http.HandlerFun
 	}
 }
 
-// Stores files in the correct location
-func (cuh *ContentUploadHandler) addFiles(files []ContentFile, contentType ContentType) error {
+func (cuh *ContentUploadHandler) handleUploadPayload(payload UploadPayload, contentType ContentType) error {
 	var contentPath string
 
 	switch contentType {
@@ -82,10 +88,11 @@ func (cuh *ContentUploadHandler) addFiles(files []ContentFile, contentType Conte
 	}
 
 	uploadedCars := make(map[string]bool)
+	uploadedTracks := make(map[string]bool)
 
 	var tags []string
 
-	for _, file := range files {
+	for _, file := range payload.Files {
 		if file.Name == "tags" {
 			tags = strings.Split(file.Data, ",")
 			continue
@@ -128,6 +135,8 @@ func (cuh *ContentUploadHandler) addFiles(files []ContentFile, contentType Conte
 
 		if contentType == ContentTypeCar {
 			uploadedCars[parts[0]] = true
+		} else if contentType == ContentTypeTrack {
+			uploadedTracks[parts[0]] = true
 		}
 
 		err = ioutil.WriteFile(path, fileDecoded, 0644)
@@ -161,6 +170,26 @@ func (cuh *ContentUploadHandler) addFiles(files []ContentFile, contentType Conte
 
 			if err != nil {
 				return err
+			}
+		}
+	}
+
+	if contentType == ContentTypeTrack && payload.GenerateTrackMaps {
+		for track := range uploadedTracks {
+			t, err := cuh.trackManager.GetTrackFromName(track)
+
+			if err != nil {
+				return err
+			}
+
+			for _, layout := range t.Layouts {
+				if layout == defaultLayoutName {
+					layout = ""
+				}
+
+				if err := cuh.trackManager.BuildTrackMap(track, layout); err != nil {
+					return err
+				}
 			}
 		}
 	}